+++ src/test/converter2/issues/gh2932.ts
@@ -1,0 +1,6 @@
+/**
+ * @inline
+ */
+type Vector2D = [start: number, end: number];
+
+export function doStuff([start, end]: Vector2D) {}